This merge is preliminary

+++ src/partition/unitquad.rs
@@ -6,7 +6,7 @@ use partition::{Partition, Subdivide};
 
 
 /// A partition of the unit quad [0, 1) × [0, 1)
-#[derive(Copy, Clone, Show, PartialEq, Hash, Eq)]
+#[derive(Copy, Clone, Debug, PartialEq, Hash, Eq)]
 pub struct UnitQuad {
     scale: u8,
     offset: (u32, u32),
@@ -18,7 +18,7 @@ impl UnitQuad {
     /// This asserts that the offset is valid given the scale level.
     pub fn new(scale: u8, offset: (u32, u32)) -> UnitQuad {
         assert!(scale < 32); // Otherwise exponentiation will overflow
-        let max_offset = 2.pow(scale as usize);
+        let max_offset = 2.pow(scale as u32);
         assert!(offset.0 < max_offset && offset.1 < max_offset);
         UnitQuad { scale: scale, offset: offset }
     }
@@ -84,14 +84,14 @@ impl Arbitrary for UnitQuad {
         // the `.gen_range` calls, even though `Rng: Gen`. The compiler
         // complains that the "source trait is private". Curiously, adding this
         // import here fixes the same situation in the `cubemap` module as well.
-        use std::rand::Rng;
+        use rand::Rng;
         let scale: u8 = {
             // scale >= 32 is invalid (overflow)
             // At scale >= 31 subdivision fails
             let max_scale = cmp::min(31, g.size()) as u8;
             g.gen_range(0, max_scale)
         };
-        let max_offset = 2.pow(scale as usize);
+        let max_offset = 2.pow(scale as u32);
         UnitQuad::new(scale, (
             g.gen_range(0, max_offset),
             g.gen_range(0, max_offset),
